+++ src/services/cliente.service.js
@@ -1,6 +1,10 @@
 import ClienteRepository from "../repositories/cliente.repository.js"
+import VendasService from "./venda.service.js";
+import bcrypt from 'bcrypt';
 
 async function createCliente(cliente) {
+        const hashPassword = await bcrypt.hash(cliente.senha, 10);
+        cliente.senha = hashPassword;
         return await ClienteRepository.insertCliente(cliente);
     
 }
@@ -14,12 +18,20 @@ async function getCliente(id) {
 }
 
 async function deleteCliente(id) {
-    //precisa verificar se não tem venda associada
+    const vendas = await VendasService.getVendasByCliente(id);
+    if (vendas.length>0) {
+        throw new Error("Há venda relacioanada ao cliente, e não pode ser excluído!");    
+    }
     await ClienteRepository.deleteCliente(id);
+    
+
+    
 }
 
 async function updateCliente(cliente)  {
     if (await ClienteRepository.getCliente(cliente.clienteId)) {    
+        const hashPassword = await bcrypt.hash(cliente.senha, 10);
+        cliente.senha = hashPassword;        
         return await ClienteRepository.updateCliente(cliente);
     }
     throw new Error("O cliente_id informado não existe");
